+++ src/theme/mod.rs
@@ -13,118 +13,74 @@ pub enum Theme {
 }
 
 impl Theme {
-  pub fn consume(self) -> AppResult<Box<dyn ColorTheme>> {
+  pub fn consume(self) -> AppResult<ColorTheme> {
     match self {
       Theme::Auto => detect_colortheme(),
-      Theme::Dark => Ok(Box::new(DarkTheme)),
-      Theme::Light => Ok(Box::new(LightTheme)),
+      Theme::Dark => Ok(dark_theme()),
+      Theme::Light => Ok(light_theme()),
     }
   }
 }
 
-fn detect_colortheme() -> AppResult<Box<dyn ColorTheme>> {
+fn detect_colortheme() -> AppResult<ColorTheme> {
   let colors = color_palette(QueryOptions::default())?;
   match colors.color_scheme() {
-    ColorScheme::Dark => Ok(Box::new(DarkTheme)),
-    ColorScheme::Light => Ok(Box::new(LightTheme)),
+    ColorScheme::Dark => Ok(dark_theme()),
+    ColorScheme::Light => Ok(light_theme()),
   }
-}
-
-pub trait ColorTheme: std::fmt::Debug {
-  fn border_color(&self) -> Color;
-  fn diff_age_color(&self) -> Color;
-  fn diff_author_color(&self) -> Color;
-  fn diff_message_color(&self) -> Color;
-  fn diff_row_hightlight_style(&self) -> Style;
-  fn diff_sha_color(&self) -> Color;
-  fn header_selected_color(&self) -> Color;
-  fn help_header_style(&self) -> Style;
-  fn help_text_style(&self) -> Style;
-  fn repo_row_hightlight_style(&self) -> Style;
-  fn text_color(&self) -> Color;
 }
 
 #[derive(Debug)]
-pub struct DarkTheme;
-#[derive(Debug)]
-pub struct LightTheme;
+pub struct ColorTheme {
+  pub border_color: Color,
+  pub diff_age_color: Color,
+  pub diff_author_color: Color,
+  pub diff_message_color: Color,
+  pub diff_row_hightlight_style: Style,
+  pub diff_sha_color: Color,
+  pub header_selected_color: Color,
+  pub help_header_style: Style,
+  pub help_text_style: Style,
+  pub repo_row_hightlight_style: Style,
+  pub text_color: Color,
+}
 
-impl ColorTheme for DarkTheme {
-  fn border_color(&self) -> Color {
-    Color::LightCyan
-  }
-  fn diff_age_color(&self) -> Color {
-    Color::Red
-  }
-  fn diff_author_color(&self) -> Color {
-    Color::Yellow
-  }
-  fn diff_message_color(&self) -> Color {
-    Color::White
-  }
-  fn diff_row_hightlight_style(&self) -> Style {
-    Style::default().add_modifier(Modifier::BOLD)
-  }
-  fn diff_sha_color(&self) -> Color {
-    Color::LightCyan
-  }
-  fn header_selected_color(&self) -> Color {
-    Color::Red
-  }
-  fn help_header_style(&self) -> Style {
-    Style::default().fg(Color::Gray).add_modifier(Modifier::DIM)
-  }
-  fn help_text_style(&self) -> Style {
-    Style::default().fg(Color::Cyan)
-  }
-  fn repo_row_hightlight_style(&self) -> Style {
-    Style::default().add_modifier(Modifier::BOLD)
-  }
-  fn text_color(&self) -> Color {
-    Color::LightCyan
+fn dark_theme() -> ColorTheme {
+  ColorTheme {
+    border_color: Color::LightCyan,
+    diff_age_color: Color::Red,
+    diff_author_color: Color::Yellow,
+    diff_message_color: Color::White,
+    diff_row_hightlight_style: Style::default().add_modifier(Modifier::BOLD),
+    diff_sha_color: Color::LightCyan,
+    header_selected_color: Color::Red,
+    help_header_style: Style::default().fg(Color::Gray).add_modifier(Modifier::DIM),
+    help_text_style: Style::default().fg(Color::Cyan),
+    repo_row_hightlight_style: Style::default().add_modifier(Modifier::BOLD),
+    text_color: Color::LightCyan,
   }
 }
 
-impl ColorTheme for LightTheme {
-  fn border_color(&self) -> Color {
-    Color::Cyan
-  }
-  fn diff_age_color(&self) -> Color {
-    Color::Red
-  }
-  fn diff_author_color(&self) -> Color {
-    Color::Blue
-  }
-  fn diff_message_color(&self) -> Color {
-    Color::Black
-  }
-  fn diff_row_hightlight_style(&self) -> Style {
-    Style::default()
+fn light_theme() -> ColorTheme {
+  ColorTheme {
+    border_color: Color::Cyan,
+    diff_age_color: Color::Red,
+    diff_author_color: Color::Blue,
+    diff_message_color: Color::Black,
+    diff_row_hightlight_style: Style::default()
       .add_modifier(Modifier::UNDERLINED)
-      .add_modifier(Modifier::BOLD)
-  }
-  fn diff_sha_color(&self) -> Color {
-    Color::Magenta
-  }
-  fn header_selected_color(&self) -> Color {
-    Color::Red
-  }
-  fn help_header_style(&self) -> Style {
-    Style::default()
+      .add_modifier(Modifier::BOLD),
+    diff_sha_color: Color::Magenta,
+    header_selected_color: Color::Red,
+    help_header_style: Style::default()
       .fg(Color::Black)
-      .add_modifier(Modifier::DIM)
-  }
-  fn help_text_style(&self) -> Style {
-    Style::default()
+      .add_modifier(Modifier::DIM),
+    help_text_style: Style::default()
       .fg(Color::Black)
-      .add_modifier(Modifier::DIM)
-  }
-  fn repo_row_hightlight_style(&self) -> Style {
-    Style::default()
+      .add_modifier(Modifier::DIM),
+    repo_row_hightlight_style: Style::default()
       .add_modifier(Modifier::UNDERLINED)
-      .add_modifier(Modifier::BOLD)
-  }
-  fn text_color(&self) -> Color {
-    Color::Black
+      .add_modifier(Modifier::BOLD),
+    text_color: Color::Black,
   }
 }