+++ src/ui/repos.rs
@@ -10,10 +10,10 @@ pub fn render(app: &App) -> Table {
   let container = Block::default()
     .title(title(app))
     .borders(Borders::ALL)
-    .style(Style::default().fg(app.theme.border_color()));
+    .style(Style::default().fg(app.theme.border_color));
 
   let rows = app.repos.iter().map(|(_id, repo)| {
-    let repo_name = text::Span::styled(repo.to_string(), Style::default().fg(app.theme.text_color()));
+    let repo_name = text::Span::styled(repo.to_string(), Style::default().fg(app.theme.text_color));
     let cells = [status_icon(repo, app), Cell::from(repo_name)];
     Row::new(cells)
   });
@@ -22,8 +22,8 @@ pub fn render(app: &App) -> Table {
 
   Table::new(rows, widths)
     .block(container)
-    .row_highlight_style(app.theme.repo_row_hightlight_style())
-    .style(Style::default().fg(app.theme.text_color()))
+    .row_highlight_style(app.theme.repo_row_hightlight_style)
+    .style(Style::default().fg(app.theme.text_color))
     .highlight_symbol("·")
     .column_spacing(2)
 }
@@ -44,5 +44,5 @@ fn status_icon<'a>(repo: &'a Repo, app: &'a App) -> Cell<'a> {
     RepoStatus::Log => " ☈",
     RepoStatus::Pulling => " ⤵",
   };
-  Cell::from(text::Span::styled(icon, Style::default().fg(app.theme.text_color())))
+  Cell::from(text::Span::styled(icon, Style::default().fg(app.theme.text_color)))
 }